Refactor from section 7 to section 2
